fixed build errors, moved old views out
