+++ .archive/views/templates/scripts.pug
