+++ package.json
@@ -13,6 +13,7 @@
     "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\""
   },
   "dependencies": {
+    "@types/reach__router": "^1.2.4",
     "gatsby": "^2.8.2",
     "node-sass": "^4.12.0",
     "react": "^16.8.6",
