+++ src/pages/index.tsx
@@ -1,16 +1,17 @@
-import React from "react"
-import { NavbarComponent } from "./components/navbar";
-import { Route } from "react-router";
-import { BrowserRouter as Router } from "react-router-dom";
-import { Home } from "./home";
+import * as React from "react"
+import NavbarComponent from "../components/navbar"
+import { Router } from "@reach/router"
+import Home from "../components/home"
 
 export default class Root extends React.Component {
-    render() {
-        return <Router>
-            <div>
-            <NavbarComponent/>
-            <Route path="/" exact component={Home} />
-            </div>
+  render() {
+    return (
+      <div>
+        <NavbarComponent />
+        <Router>
+          <Home path="/" />
         </Router>
-    }
+      </div>
+    )
+  }
 }
