+++ tsconfig.json
@@ -2,5 +2,6 @@
 	"compilerOptions": {
 		"lib": ["dom", "es6"],
 		"target": "es3",
-	}
+		"jsx": "react"
+	},
 }
